Add namespaced filter service and test cases
